+++ apps/web/@/constants/upload.ts
@@ -1,0 +1,6 @@
+export const OrderByField = {
+  newest: "newest",
+  oldest: "oldest",
+  nameAsc: "name_asc",
+  nameDesc: "name_desc",
+}
